requirements.txt
PyYAML>=6.0 
colorama>=0.4.4